changed folder structure to modular folder structure
