+++ src/app/core/redux/app/types.ts
