+++ src/app/core/redux/store.ts
@@ -1,6 +1,6 @@
 import { createStore, combineReducers, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
-import { authReducer } from "./auth/reducer";
+import { authReducer } from "../../../auth/core/redux/reducer";
 import { composeWithDevTools } from "redux-devtools-extension";
 const middleware = [thunk];
 const reducers = combineReducers({ auth: authReducer });
