+++ src/auth/containers/forgotPassword/ForgotPassword.tsx
@@ -1,27 +1,23 @@
 import React, { useEffect } from "react";
 import { Typography } from "@material-ui/core";
-import { APP_NAME } from "../../core/constants";
+import { APP_NAME } from "../../../app/core/constants";
 
 import { connect } from "react-redux";
 
 import { Grid } from "@material-ui/core";
 
 import { useDispatch } from "react-redux";
-import { SetAuthenticated } from "../../core/store/app/actions";
-
-import ResetPasswordForm from "../../components/forgotPassword/ResetPasswordForm";
-const mainImage = require("../../assets/images/main-image.jpg");
+import { SetAuthenticated } from "../../../app/core/redux/app/actions";
+import ForgotPasswordForm from "../../components/forgotPassword/ForgotPasswordForm";
+const mainImage = require("../../../assets/images/main-image.jpg");
 
 const mapStateToProps = (state: any) => {
   return {
     state: state,
   };
 };
-const qs = require("query-string");
 
-function ResetPassword(props: any) {
-  const data = qs.parse(props.location.search);
-
+const ForgotPassword = (props: any) => {
   const dispatch = useDispatch();
 
   const { isAuthenticated } = props.state.auth;
@@ -34,8 +30,6 @@ function ResetPassword(props: any) {
       dispatch(SetAuthenticated());
     }
   });
-
-  const { token } = data;
 
   return (
     <div>
@@ -56,11 +50,11 @@ function ResetPassword(props: any) {
             {APP_NAME}
           </Typography>
           <br /> <br />
-          <ResetPasswordForm token={token} props={props} />
+          <ForgotPasswordForm props={props} />
         </Grid>
       </Grid>
     </div>
   );
-}
+};
 
-export default connect(mapStateToProps)(ResetPassword);
+export default connect(mapStateToProps)(ForgotPassword);
