+++ src/auth/containers/forgotPassword/ResetPassword.tsx
@@ -1,26 +1,28 @@
 import React, { useEffect } from "react";
 import { Typography } from "@material-ui/core";
-import { APP_NAME } from "../../core/constants";
+import { APP_NAME } from "../../../app/core/constants";
 
 import { connect } from "react-redux";
 
 import { Grid } from "@material-ui/core";
 
 import { useDispatch } from "react-redux";
-import { SetAuthenticated } from "../../core/store/app/actions";
-import RegisterForm from "../../components/register/RegisterForm";
-const mainImage = require("../../assets/images/main-image.jpg");
+import { SetAuthenticated } from "../../../app/core/redux/app/actions";
+
+import ResetPasswordForm from "../../components/forgotPassword/ResetPasswordForm";
+const mainImage = require("../../../assets/images/main-image.jpg");
 
 const mapStateToProps = (state: any) => {
   return {
     state: state,
   };
 };
+const qs = require("query-string");
 
-const Register = (props: any) => {
+function ResetPassword(props: any) {
+  const data = qs.parse(props.location.search);
+
   const dispatch = useDispatch();
-  // const message = props.state.app.message;
-  console.log("props data ", props);
 
   const { isAuthenticated } = props.state.auth;
 
@@ -32,6 +34,8 @@ const Register = (props: any) => {
       dispatch(SetAuthenticated());
     }
   });
+
+  const { token } = data;
 
   return (
     <div>
@@ -51,12 +55,12 @@ const Register = (props: any) => {
             {" "}
             {APP_NAME}
           </Typography>
-          <br />
-          <RegisterForm props={props} />
+          <br /> <br />
+          <ResetPasswordForm token={token} props={props} />
         </Grid>
       </Grid>
     </div>
   );
-};
+}
 
-export default connect(mapStateToProps)(Register);
+export default connect(mapStateToProps)(ResetPassword);
