+++ src/auth/containers/login/Login.tsx
@@ -1,15 +1,16 @@
 import React, { useEffect } from "react";
 import { Typography } from "@material-ui/core";
-import { APP_NAME } from "../../core/constants";
+import { APP_NAME } from "../../../app/core/constants";
 
 import { connect } from "react-redux";
+import "./login.css";
 
+import LoginForm from "../../components/login/LoginForm";
 import { Grid } from "@material-ui/core";
 
 import { useDispatch } from "react-redux";
-import { SetAuthenticated } from "../../core/store/app/actions";
-import ForgotPasswordForm from "../../components/forgotPassword/ForgotPasswordForm";
-const mainImage = require("../../assets/images/main-image.jpg");
+import { SetAuthenticated } from "../../../app/core/redux/app/actions";
+const mainImage = require("../../../assets/images/main-image.jpg");
 
 const mapStateToProps = (state: any) => {
   return {
@@ -17,7 +18,7 @@ const mapStateToProps = (state: any) => {
   };
 };
 
-const ForgotPassword = (props: any) => {
+const Login = (props: any) => {
   const dispatch = useDispatch();
 
   const { isAuthenticated } = props.state.auth;
@@ -34,6 +35,7 @@ const ForgotPassword = (props: any) => {
   return (
     <div>
       <Grid container spacing={0}>
+        {" "}
         <Grid item xs={1} sm={6}>
           <div className="fill">
             <img
@@ -43,18 +45,17 @@ const ForgotPassword = (props: any) => {
             />
           </div>
         </Grid>
-
-        <Grid item xs={11} sm={6}>
+        <Grid item xs={11} sm={6} color="red">
           <Typography variant="h3" color="primary" component="h2">
             {" "}
             {APP_NAME}
           </Typography>
           <br /> <br />
-          <ForgotPasswordForm props={props} />
+          <LoginForm props={props} />
         </Grid>
       </Grid>
     </div>
   );
 };
 
-export default connect(mapStateToProps)(ForgotPassword);
+export default connect(mapStateToProps)(Login);
