+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import { Provider } from "react-redux";
-import { store } from "./core/store/store";
+import { store } from "./app/core/redux/store";
 
 ReactDOM.render(
   <React.StrictMode>
