+++ src/shared/assets/images/main-image.jpg
